+++ src/components/CytoscapeGraph/MiniGraphCard.tsx
@@ -1,0 +1,124 @@
+import * as React from 'react';
+import { style } from 'typestyle';
+import {
+  Card,
+  CardActions,
+  CardBody,
+  CardHead,
+  CardHeader,
+  Dropdown,
+  DropdownItem,
+  KebabToggle,
+  Title
+} from '@patternfly/react-core';
+import history from '../../app/History';
+import GraphDataSource from '../../services/GraphDataSource';
+import { EdgeLabelMode, GraphType, NodeType } from '../../types/Graph';
+import CytoscapeGraph from './CytoscapeGraph';
+import { CytoscapeGraphSelectorBuilder } from './CytoscapeGraphSelector';
+import { DagreGraph } from './graphs/DagreGraph';
+
+const miniGraphContainerStyle = style({ height: '300px' });
+
+type MiniGraphCardProps = {
+  dataSource: GraphDataSource;
+};
+
+type MiniGraphCardState = {
+  isKebabOpen: boolean;
+};
+
+export default class MiniGraphCard extends React.Component<MiniGraphCardProps, MiniGraphCardState> {
+  constructor(props) {
+    super(props);
+    this.state = { isKebabOpen: false };
+  }
+
+  render() {
+    const graphCardActions = [
+      <DropdownItem key="viewGraph" onClick={this.onViewGraph}>
+        View full graph
+      </DropdownItem>
+    ];
+
+    return (
+      <Card style={{ height: '100%' }}>
+        <CardHead>
+          <CardActions>
+            <Dropdown
+              toggle={<KebabToggle onToggle={this.onGraphActionsToggle} />}
+              dropdownItems={graphCardActions}
+              isPlain
+              isOpen={this.state.isKebabOpen}
+              position={'right'}
+            />
+          </CardActions>
+          <CardHeader>
+            <Title style={{ float: 'left' }} headingLevel="h3" size="2xl">
+              Graph Overview
+            </Title>
+          </CardHeader>
+        </CardHead>
+        <CardBody>
+          <div style={{ height: '100%' }}>
+            <CytoscapeGraph
+              containerClassName={miniGraphContainerStyle}
+              dataSource={this.props.dataSource}
+              displayUnusedNodes={() => undefined}
+              edgeLabelMode={EdgeLabelMode.NONE}
+              isMTLSEnabled={false}
+              isMiniGraph={true}
+              layout={DagreGraph.getLayout()}
+              refreshInterval={0}
+              showCircuitBreakers={false}
+              showMissingSidecars={true}
+              showNodeLabels={true}
+              showSecurity={false}
+              showServiceNodes={true}
+              showTrafficAnimation={true}
+              showUnusedNodes={false}
+              showVirtualServices={true}
+            />
+          </div>
+        </CardBody>
+      </Card>
+    );
+  }
+
+  private onGraphActionsToggle = (isOpen: boolean) => {
+    this.setState({
+      isKebabOpen: isOpen
+    });
+  };
+
+  private onViewGraph = () => {
+    const namespace = this.props.dataSource.fetchParameters.namespaces[0].name;
+    let cytoscapeGraph = new CytoscapeGraphSelectorBuilder().namespace(namespace);
+    let graphType: GraphType = GraphType.APP;
+
+    switch (this.props.dataSource.fetchParameters.node!.nodeType) {
+      case NodeType.WORKLOAD:
+        graphType = GraphType.WORKLOAD;
+        cytoscapeGraph = cytoscapeGraph.workload(this.props.dataSource.fetchParameters.node!.workload);
+        break;
+
+      case NodeType.APP:
+        cytoscapeGraph = cytoscapeGraph
+          .app(this.props.dataSource.fetchParameters.node!.app)
+          .nodeType(NodeType.APP)
+          .isGroup(null);
+        break;
+
+      case NodeType.SERVICE:
+        graphType = GraphType.SERVICE;
+        cytoscapeGraph = cytoscapeGraph.service(this.props.dataSource.fetchParameters.node!.service);
+        break;
+    }
+
+    const graphUrl = `/graph/namespaces?graphType=${graphType}&injectServiceNodes=true&namespaces=${namespace}&unusedNodes=true&focusSelector=${encodeURI(
+      cytoscapeGraph.build()
+    )}`;
+
+    history.push(graphUrl);
+  };
+}
