+++ src/pages/ServiceDetails/ServiceInfo/ServiceInfoDescription.tsx
@@ -1,15 +1,9 @@
 import * as React from 'react';
 import {
   Card,
-  CardActions,
   CardBody,
-  CardHead,
-  CardHeader,
-  Dropdown,
-  DropdownItem,
   Grid,
   GridItem,
-  KebabToggle,
   Stack,
   StackItem,
   Text,
@@ -19,25 +13,21 @@ import {
 } from '@patternfly/react-core';
 import { EyeIcon } from '@patternfly/react-icons';
 import { style } from 'typestyle';
-import CytoscapeGraph from '../../../components/CytoscapeGraph/CytoscapeGraph';
-import { DagreGraph } from '../../../components/CytoscapeGraph/graphs/DagreGraph';
 import LocalTime from '../../../components/Time/LocalTime';
 import { DisplayMode, HealthIndicator } from '../../../components/Health/HealthIndicator';
 import GraphDataSource from '../../../services/GraphDataSource';
 import { ServiceHealth } from '../../../types/Health';
 import { Endpoints } from '../../../types/ServiceInfo';
 import { ObjectCheck, ObjectValidation, Port } from '../../../types/IstioObjects';
-import { EdgeLabelMode, GraphType } from '../../../types/Graph';
-import { CytoscapeGraphSelectorBuilder } from '../../../components/CytoscapeGraph/CytoscapeGraphSelector';
 import { ValidationObjectSummary } from '../../../components/Validations/ValidationObjectSummary';
 import ValidationList from '../../../components/Validations/ValidationList';
-import history from '../../../app/History';
 import Labels from '../../../components/Label/Labels';
 import { ThreeScaleServiceRule } from '../../../types/ThreeScale';
 import { AdditionalItem } from 'types/Workload';
 import { TextOrLink } from 'components/TextOrLink';
 import { renderAPILogo } from 'components/Logo/Logos';
 import './ServiceInfoDescription.css';
+import MiniGraphCard from '../../../components/CytoscapeGraph/MiniGraphCard';
 
 interface ServiceInfoDescriptionProps {
   name: string;
@@ -70,8 +60,6 @@ const listStyle = style({
 
 const ExternalNameType = 'ExternalName';
 
-const cytoscapeGraphContainerStyle = style({ height: '300px' });
-
 class ServiceInfoDescription extends React.Component<ServiceInfoDescriptionProps, ServiceInfoDescriptionState> {
   constructor(props: ServiceInfoDescriptionProps) {
     super(props);
@@ -92,12 +80,6 @@ class ServiceInfoDescription extends React.Component<ServiceInfoDescriptionProps
   }
 
   render() {
-    const graphCardActions = [
-      <DropdownItem key="viewGraph" onClick={this.onViewGraph}>
-        View full graph
-      </DropdownItem>
-    ];
-
     return (
       <Grid gutter="md">
         <GridItem span={4}>
@@ -144,48 +126,7 @@ class ServiceInfoDescription extends React.Component<ServiceInfoDescriptionProps
           </Card>
         </GridItem>
         <GridItem span={4}>
-          <Card style={{ height: '100%' }}>
-            <CardHead>
-              <CardActions>
-                <Dropdown
-                  toggle={<KebabToggle onToggle={this.onGraphActionsToggle} />}
-                  dropdownItems={graphCardActions}
-                  isPlain
-                  isOpen={this.state.isGraphActionsOpen}
-                  position={'right'}
-                />
-              </CardActions>
-              <CardHeader>
-                <Title headingLevel="h3" size="2xl">
-                  Graph Overview
-                </Title>
-              </CardHeader>
-            </CardHead>
-            <CardBody>
-              <div style={{ height: '100%' }}>
-                <CytoscapeGraph
-                  activeNamespaces={[{ name: this.props.namespace }]}
-                  containerClassName={cytoscapeGraphContainerStyle}
-                  dataSource={this.props.miniGraphDatasource}
-                  displayUnusedNodes={() => undefined}
-                  edgeLabelMode={EdgeLabelMode.NONE}
-                  graphType={GraphType.APP}
-                  isMTLSEnabled={false}
-                  isMiniGraph={true}
-                  layout={DagreGraph.getLayout()}
-                  refreshInterval={0}
-                  showCircuitBreakers={false}
-                  showMissingSidecars={true}
-                  showNodeLabels={true}
-                  showSecurity={false}
-                  showServiceNodes={true}
-                  showTrafficAnimation={true}
-                  showUnusedNodes={false}
-                  showVirtualServices={true}
-                />
-              </div>
-            </CardBody>
-          </Card>
+          <MiniGraphCard dataSource={this.props.miniGraphDatasource} />
         </GridItem>
         <GridItem span={4}>
           <Card style={{ height: '100%' }}>
@@ -265,22 +206,6 @@ class ServiceInfoDescription extends React.Component<ServiceInfoDescriptionProps
       </Grid>
     );
   }
-
-  private onGraphActionsToggle = (isOpen: boolean) => {
-    this.setState({
-      isGraphActionsOpen: isOpen
-    });
-  };
-
-  private onViewGraph = () => {
-    let cytoscapeGraph = new CytoscapeGraphSelectorBuilder().namespace(this.props.namespace).service(this.props.name);
-
-    const graphUrl = `/graph/namespaces?graphType=${GraphType.SERVICE}&injectServiceNodes=true&namespaces=${
-      this.props.namespace
-    }&unusedNodes=true&focusSelector=${encodeURI(cytoscapeGraph.build())}`;
-
-    history.push(graphUrl);
-  };
 }
 
 export default ServiceInfoDescription;
