+++ src/pages/WorkloadDetails/WorkloadInfo/WorkloadDescription.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { style } from 'typestyle';
 import { Workload } from '../../../types/Workload';
 import LocalTime from '../../../components/Time/LocalTime';
 import { DisplayMode, HealthIndicator } from '../../../components/Health/HealthIndicator';
@@ -7,15 +6,9 @@ import { WorkloadHealth } from '../../../types/Health';
 import Labels from '../../../components/Label/Labels';
 import {
   Card,
-  CardActions,
   CardBody,
-  CardHead,
-  CardHeader,
-  Dropdown,
-  DropdownItem,
   Grid,
   GridItem,
-  KebabToggle,
   PopoverPosition,
   Stack,
   StackItem,
@@ -25,14 +18,8 @@ import {
 } from '@patternfly/react-core';
 import { TextOrLink } from 'components/TextOrLink';
 import { renderRuntimeLogo, renderAPILogo } from 'components/Logo/Logos';
-import history from '../../../app/History';
-import CytoscapeGraph from '../../../components/CytoscapeGraph/CytoscapeGraph';
-import { CytoscapeGraphSelectorBuilder } from '../../../components/CytoscapeGraph/CytoscapeGraphSelector';
-import { DagreGraph } from '../../../components/CytoscapeGraph/graphs/DagreGraph';
-import { EdgeLabelMode, GraphType } from '../../../types/Graph';
 import GraphDataSource from '../../../services/GraphDataSource';
-
-const cytoscapeGraphContainerStyle = style({ height: '300px' });
+import MiniGraphCard from '../../../components/CytoscapeGraph/MiniGraphCard';
 
 type WorkloadDescriptionProps = {
   workload: Workload;
@@ -42,27 +29,13 @@ type WorkloadDescriptionProps = {
   miniGraphDataSource: GraphDataSource;
 };
 
-type WorkloadDescriptionState = {
-  isGraphActionsOpen: boolean;
-};
-
-class WorkloadDescription extends React.Component<WorkloadDescriptionProps, WorkloadDescriptionState> {
-  constructor(props: WorkloadDescriptionProps) {
-    super(props);
-    this.state = { isGraphActionsOpen: false };
-  }
-
+class WorkloadDescription extends React.Component<WorkloadDescriptionProps> {
   render() {
     const workload = this.props.workload;
     const isTemplateLabels =
       workload &&
       ['Deployment', 'ReplicaSet', 'ReplicationController', 'DeploymentConfig', 'StatefulSet'].indexOf(workload.type) >=
         0;
-    const graphCardActions = [
-      <DropdownItem key="viewGraph" onClick={this.onViewGraph}>
-        View full graph
-      </DropdownItem>
-    ];
     const runtimes = workload.runtimes.map(r => r.name).filter(name => name !== '');
     return workload ? (
       <Grid gutter="md">
@@ -116,48 +89,7 @@ class WorkloadDescription extends React.Component<WorkloadDescriptionProps, Work
           </Card>
         </GridItem>
         <GridItem span={4}>
-          <Card style={{ height: '100%' }}>
-            <CardHead>
-              <CardActions>
-                <Dropdown
-                  toggle={<KebabToggle onToggle={this.onGraphActionsToggle} />}
-                  dropdownItems={graphCardActions}
-                  isPlain
-                  isOpen={this.state.isGraphActionsOpen}
-                  position={'right'}
-                />
-              </CardActions>
-              <CardHeader>
-                <Title headingLevel="h3" size="2xl">
-                  Graph Overview{' '}
-                </Title>
-              </CardHeader>
-            </CardHead>
-            <CardBody>
-              <div style={{ height: '300px' }}>
-                <CytoscapeGraph
-                  activeNamespaces={[{ name: this.props.namespace }]}
-                  containerClassName={cytoscapeGraphContainerStyle}
-                  dataSource={this.props.miniGraphDataSource}
-                  displayUnusedNodes={() => undefined}
-                  edgeLabelMode={EdgeLabelMode.NONE}
-                  graphType={GraphType.APP}
-                  isMTLSEnabled={false}
-                  isMiniGraph={true}
-                  layout={DagreGraph.getLayout()}
-                  refreshInterval={0}
-                  showCircuitBreakers={false}
-                  showMissingSidecars={true}
-                  showNodeLabels={true}
-                  showSecurity={false}
-                  showServiceNodes={true}
-                  showTrafficAnimation={true}
-                  showUnusedNodes={false}
-                  showVirtualServices={true}
-                />
-              </div>
-            </CardBody>
-          </Card>
+          <MiniGraphCard dataSource={this.props.miniGraphDataSource} />
         </GridItem>
         <GridItem span={4}>
           <Card style={{ height: '100%' }}>
@@ -185,24 +117,6 @@ class WorkloadDescription extends React.Component<WorkloadDescriptionProps, Work
       'Loading'
     );
   }
-
-  private onGraphActionsToggle = (isOpen: boolean) => {
-    this.setState({
-      isGraphActionsOpen: isOpen
-    });
-  };
-
-  private onViewGraph = () => {
-    let cytoscapeGraph = new CytoscapeGraphSelectorBuilder()
-      .namespace(this.props.namespace)
-      .workload(this.props.workload.name);
-
-    const graphUrl = `/graph/namespaces?graphType=${GraphType.WORKLOAD}&injectServiceNodes=true&namespaces=${
-      this.props.namespace
-    }&unusedNodes=true&focusSelector=${encodeURI(cytoscapeGraph.build())}`;
-
-    history.push(graphUrl);
-  };
 }
 
 export default WorkloadDescription;
